# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Portfolio Manager API",
    "description": "Phase 1 backend. Not financial advice.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/securities/search": {
      "get": {
        "tags": [
          "misc"
        ],
        "summary": "Securities Search",
        "description": "Seeded or provider-verified securities only (never another user's unverified ticker).",
        "operationId": "securities_search_api_securities_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 64,
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SecurityHit"
                  },
                  "title": "Response Securities Search Api Securities Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/holdings/{holding_id}/scorecard": {
      "get": {
        "tags": [
          "misc"
        ],
        "summary": "Scorecard",
        "operationId": "scorecard_api_holdings__holding_id__scorecard_get",
        "parameters": [
          {
            "name": "holding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Holding Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreCardDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/alerts": {
      "get": {
        "tags": [
          "misc"
        ],
        "summary": "Alerts",
        "operationId": "alerts_api_alerts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AlertOut"
                  },
                  "type": "array",
                  "title": "Response Alerts Api Alerts Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "misc"
        ],
        "summary": "Create Alert",
        "operationId": "create_alert_api_alerts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/alerts/{alert_id}": {
      "delete": {
        "tags": [
          "misc"
        ],
        "summary": "Delete Alert",
        "operationId": "delete_alert_api_alerts__alert_id__delete",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Alert Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/notifications/{notification_id}/read": {
      "post": {
        "tags": [
          "misc"
        ],
        "summary": "Read Notification",
        "operationId": "read_notification_api_notifications__notification_id__read_post",
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Notification Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AlertCreate": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "op": {
            "type": "string",
            "enum": [
              "above",
              "below"
            ],
            "title": "Op"
          },
          "price": {
            "type": "number",
            "maximum": 1000000000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Price"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "symbol",
          "op",
          "price"
        ],
        "title": "AlertCreate"
      },
      "AlertOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "op": {
            "type": "string",
            "enum": [
              "above",
              "below"
            ],
            "title": "Op"
          },
          "price": {
            "type": "number",
            "title": "Price"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "triggered_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Triggered At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "symbol",
          "op",
          "price",
          "active"
        ],
        "title": "AlertOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NotificationOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "read": {
            "type": "boolean",
            "title": "Read"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "title",
          "body",
          "created_at",
          "read"
        ],
        "title": "NotificationOut"
      },
...
      "ScoreCardDetail": {
        "properties": {
          "holding_id": {
            "type": "integer",
            "title": "Holding Id"
          },
          "portfolio_id": {
            "type": "integer",
            "title": "Portfolio Id"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "name_en": {
            "type": "string",
            "title": "Name En"
          },
          "name_he": {
            "type": "string",
            "title": "Name He"
          },
          "horizon": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "1w",
                  "1m",
                  "3m",
                  "6m",
                  "1y"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Horizon"
          },
          "total": {
            "type": "number",
            "title": "Total"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "validated": {
            "type": "boolean",
            "title": "Validated"
          },
          "signals": {
            "items": {
              "$ref": "#/components/schemas/SignalBreakdownOut"
            },
            "type": "array",
            "title": "Signals"
          },
          "explanation": {
            "additionalProperties": true,
            "type": "object",
            "title": "Explanation"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer"
          }
        },
        "type": "object",
        "required": [
          "holding_id",
          "portfolio_id",
          "symbol",
          "name_en",
          "name_he",
          "total",
          "confidence",
          "available",
          "validated",
          "signals",
          "explanation",
          "disclaimer"
        ],
        "title": "ScoreCardDetail"
      },
...
      "SecurityHit": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "name_en": {
            "type": "string",
            "title": "Name En"
          },
          "name_he": {
            "type": "string",
            "title": "Name He"
          },
          "market": {
            "type": "string",
            "enum": [
              "US",
              "TASE",
              "CRYPTO"
            ],
            "title": "Market"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "name_en",
          "name_he",
          "market"
        ],
        "title": "SecurityHit"
      },
...
      "SignalBreakdownOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "weight": {
            "type": "number",
            "title": "Weight"
          },
          "nominal_weight": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nominal Weight"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          },
          "data_as_of": {
            "type": "string",
            "title": "Data As Of"
          },
          "explanation": {
            "additionalProperties": true,
            "type": "object",
            "title": "Explanation"
          }
        },
        "type": "object",
        "required": [
          "name",
          "score",
          "confidence",
          "weight",
          "reasons",
          "data_as_of",
          "explanation"
        ],
        "title": "SignalBreakdownOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}